Extract GraphEdit node arrangement
